% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add_cmake.R
\name{add_cmake}
\alias{add_cmake}
\title{Add cmake files to the R package.}
\usage{
add_cmake(dir, project, language = c("CXX", "C", "Fortran"),
  cxx_standard = c("11", "14", "98"), compileAttributes = TRUE)
}
\arguments{
\item{dir}{the path of R package.}

\item{project}{project name to be created in cmake. Retrieve from
DESCRIPTION file if it is not set.}

\item{language}{programming lagnuage to be compiled.}

\item{cxx_standard}{cxx_standard used.}
}
\description{
Add cmake files to the R package.
}
\author{
Chenliang Xu
}

