% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cmake.R
\name{cmake}
\alias{cmake}
\title{Run cmake within R}
\usage{
cmake(options = "--version", wd = getwd(), env_vars = NULL,
  cmake_command = cmaker_command_path("cmake"), ...)
}
\arguments{
\item{options}{a character vector of options pass to cmake.}

\item{wd}{the work directory to run cmake in.}

\item{env_vars}{environment variables to set before running the command.}

\item{cmake_command}{the name or the path of cmake command in the system.}

\item{...}{additional arguments passed to
\code{\link{devtools::system_check}}}
}
\value{
Invisible \code{TRUE} if the command succeeds, throws an error if
the command fails.
}
\description{
Run cmake within R
}
\author{
Chenliang Xu
}

