% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flags.R
\name{cmakeCommit}
\alias{cmakeCommit}
\title{return cmake commit information}
\usage{
cmakeCommit()
}
\description{
call this to get the installed cmake commit
\url{https://github.com/Kitware/CMake/}
}
\examples{
cmakeCommit()
}
\author{
Avants BB
}

