% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/command_path.R
\name{cmaker_command_path}
\alias{cmaker_command_path}
\title{Get the system command path}
\usage{
cmaker_command_path(command)
}
\arguments{
\item{command}{command name}
}
\value{
path of command
}
\description{
Retrive the path from option("cmaker.command.path") if exists;
otherwise find command with Sys.which
}
\author{
Chenliang Xu
}

