% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flags.R
\name{cmakeVersion}
\alias{cmakeVersion}
\title{return cmake version information}
\usage{
cmakeVersion()
}
\description{
call this to get the installed cmake version
}
\examples{
cmakeVersion()
}
\author{
Avants BB
}

