% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/command_exist.R
\name{command_exist}
\alias{command_exist}
\title{Test existence of a system command.}
\usage{
command_exist(command, test_options)
}
\arguments{
\item{command}{The name or path of command to run.}

\item{test_options}{The options passed to command for testing run.}
}
\value{
TRUE if the command successes; FALSE if it fails.
}
\description{
Test existence of a system command by trying to run the command
with test_options.
}
\author{
Chenliang Xu
}

