% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/include_path.R
\name{include_path}
\alias{include_path}
\title{Include path for compiling the package.}
\usage{
include_path(dir = ".")
}
\arguments{
\item{dir}{The directory of package.}
}
\value{
The include path of package.
}
\description{
The implementation is copied from `tools:::.install_packages` and
modifyed gently.
}
\author{
Chenliang Xu
}

