% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_asis.R
\name{load_asis}
\alias{load_asis}
\title{Load the package asis.}
\usage{
load_asis(pkg = ".", reset = TRUE, export_all = TRUE)
}
\arguments{
\item{pkg}{package description, can be path or package name.}

\item{reset}{clear package environment and reset file cache before loading
any pieces of the package. This is equivalent to running
\code{\link{unload}} and is the default. Use \code{reset = FALSE} may be
faster for large code bases, but is a significantly less accurate
approximation.}

\item{export_all}{If \code{TRUE} (the default), export all objects.
If \code{FALSE}, export only the objects that are listed as exports
in the NAMESPACE file.}
}
\description{
Load the package asis.
}
\details{
Load a package asis. It's similar to `devtools::load_all` but
never compile a DLL. Instead, it assumes the DLL has been compiled
properly. See `devtools::load_all` for other detials.

The implementation is copied from `devtools::load_all` and
modifyed gently.
}
\author{
Chenliang Xu
}

