% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generate_project.R
\name{ls_generators}
\alias{ls_generators}
\title{Get supported cmake generators.}
\usage{
ls_generators()
}
\value{
supported generators
}
\description{
Get supported cmake generators.
}
\author{
Chenliang Xu
}

