% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkg_name.R
\name{pkg_name}
\alias{pkg_name}
\title{R package name.}
\usage{
pkg_name(dir = ".")
}
\arguments{
\item{dir}{The directory of package.}
}
\value{
The package name.
}
\description{
Get the R package name.
}
\details{
Parse the DESCRIPTION file, and retrive the R package name.
}
\author{
Chenliang Xu
}

