% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/template.R
\name{template}
\alias{template}
\title{Simple template engine.}
\usage{
template(template, settings)
}
\arguments{
\item{template}{a character verctor storing the template.}

\item{settings}{the setting variables to be repalced.}
}
\value{
the randered tempalte.
}
\description{
A simple template engine that replace "@var_name@" by values
from settings.
}
\author{
Chenliang Xu
}

