% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{cmake}
\alias{cmake}
\alias{cmake_path}
\title{return cmake binary location}
\usage{
cmake()

cmake_path()
}
\description{
call this to get the cmake binary location to help you build an R package
}
\examples{

cmake()

}
\author{
Avants BB
}
