% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flags.R
\name{cmake}
\alias{cmake}
\title{return cmake binary location}
\usage{
cmake()
}
\description{
call this to get the cmake binary location to help you build an R package
}
\examples{
cmake()
}
\author{
Avants BB
}

