% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generate_project.R
\name{generate_project}
\alias{generate_project}
\title{Generate project file for IDE.}
\usage{
generate_project(dir, generator, ...)
}
\arguments{
\item{dir}{the directory}

\item{generator}{the cmake generator to use}

\item{...}{args pass to cmake}
}
\description{
Call cmake to generate project file for IDE.
}
\author{
Chenliang Xu
}

