% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_project.R
\name{generate_project}
\alias{generate_project}
\title{Generate project file for IDE.}
\usage{
generate_project(dir, IDE, cmake_options = "")
}
\arguments{
\item{dir}{the directory of R package}

\item{IDE}{the IDE to generate. It's the same as cmake generator.}

\item{cmake_options}{additon cmake options.}
}
\description{
Generate project file for IDE. It calles `cmake` to generate
project file, and then adjust the generated files if needed.
}
\author{
Chenliang Xu
}

