% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list_sources.R
\name{list_sources}
\alias{list_sources}
\title{List source files of package}
\usage{
list_sources(dir, out, collapse = "\\n", on_change = NULL)
}
\arguments{
\item{dir}{the directory of package}

\item{out}{Set to logical true to print resutls on screen; set a
character file name, to print result to the file.}

\item{collapse}{character to separate results for printing}
}
\value{
An invisible list of source file names
}
\description{
List source files of package
}
\author{
Chenliang Xu
}

